better naming of parts
