+++ electronics_lib/test_inductor.py
@@ -7,7 +7,7 @@ from .test_passive_common import *
 class InductorTestTop(Block):
   def __init__(self):
     super().__init__()
-    self.dut = self.Block(SmtInductor(
+    self.dut = self.Block(DigikeyInductor(
       inductance=2.2 * uHenry(tol=0.2),
       frequency=(0, 1) * MHertz,
       current=(0, 500) * mAmp
